model fastapi endpoint/requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
numpy
scikit-image
python-multipart
